# openapi/v1.yaml
openapi: 3.0.3
info:
  title: ShiftLush API
  version: v1
paths:
...
  /api/v1/elevators/:
    get:
      operationId: elevators_list
      description: |-
        CRUD over a company-owned model.

        PUT is not routed: partial update is the only update verb. A full replace
        means the client has to send every field it does not want cleared, and the
        moment the model grows a column, old clients start wiping it.
      parameters:
      - in: query
        name: building
        schema:
          type: string
          format: uuid
      - in: query
        name: category
        schema:
          type: string
          enum:
          - accessibility_platform
          - dumbwaiter
          - freight
          - passenger
          - passenger_freight
          - vehicle
        description: |-
          * `passenger` - Passenger
          * `freight` - Freight
          * `passenger_freight` - Passenger and freight
          * `dumbwaiter` - Dumbwaiter
          * `accessibility_platform` - Accessibility platform
          * `vehicle` - Vehicle
      - in: query
        name: customer
        schema:
          type: string
          format: uuid
      - in: query
        name: inspection_label
        schema:
          type: string
          enum:
          - blue
          - green
          - none
          - red
          - yellow
        description: |-
          * `green` - Green
          * `blue` - Blue
          * `yellow` - Yellow
          * `red` - Red
          * `none` - None
      - name: ordering
        required: false
        in: query
        description: Sonuçların sıralanmasında kullanılacak alan.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Sayfalanmış sonuç kümesinde bir sayfa numarası.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Her sayfada döndürülecek sonuç sayısı.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
          enum:
          - active
          - out_of_service
          - sealed
          - suspended
          - uncontracted
        description: |-
          * `active` - Active
          * `suspended` - Suspended
          * `sealed` - Sealed
          * `out_of_service` - Out of service
          * `uncontracted` - Uncontracted
      tags:
      - elevators
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedElevatorListList'
          description: ''
    post:
      operationId: elevators_create
      description: |-
        CRUD over a company-owned model.

        PUT is not routed: partial update is the only update verb. A full replace
        means the client has to send every field it does not want cleared, and the
        moment the model grows a column, old clients start wiping it.
      tags:
      - elevators
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElevatorWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ElevatorWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ElevatorWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElevatorWrite'
          description: ''
...
components:
  schemas:
...
    BlankEnum:
      enum:
      - ''
...
    ControlTypeEnum:
      enum:
      - simple_collective
      - down_collective
      - full_collective
      - group_control
      type: string
      description: |-
        * `simple_collective` - Simple collective
        * `down_collective` - Down collective
        * `full_collective` - Full collective
        * `group_control` - Group control
...
    DoorTypeEnum:
      enum:
      - automatic_center
      - automatic_side
      - semi_automatic
      - manual
      type: string
      description: |-
        * `automatic_center` - Automatic centre opening
        * `automatic_side` - Automatic side opening
        * `semi_automatic` - Semi automatic
        * `manual` - Manual
    DriveTypeEnum:
      enum:
      - geared_electric
      - gearless_electric
      - hydraulic
      type: string
      description: |-
        * `geared_electric` - Geared electric
        * `gearless_electric` - Gearless electric
        * `hydraulic` - Hydraulic
...
    ElevatorCategory:
      enum:
      - passenger
      - freight
      - passenger_freight
      - dumbwaiter
      - accessibility_platform
      - vehicle
      type: string
      description: |-
        * `passenger` - Passenger
        * `freight` - Freight
        * `passenger_freight` - Passenger and freight
        * `dumbwaiter` - Dumbwaiter
        * `accessibility_platform` - Accessibility platform
        * `vehicle` - Vehicle
...
    ElevatorList:
      type: object
      description: |-
        What the list screen needs — seven columns' worth, not all 31 fields.

        The rest are record fields rather than search criteria and live on the
        detail response; sending them all would make the heaviest screen in the
        product heavier for nothing.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        registration_number:
          type: string
          readOnly: true
        name:
          type: string
          readOnly: true
        category:
          allOf:
          - $ref: '#/components/schemas/ElevatorCategory'
          readOnly: true
        stop_count:
          type: integer
          readOnly: true
          nullable: true
        has_car_door:
          type: boolean
          readOnly: true
          nullable: true
        building_id:
          type: string
          format: uuid
          readOnly: true
        building_name:
          type: string
          readOnly: true
        customer_name:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ElevatorStatus'
          readOnly: true
        inspection_label:
          allOf:
          - $ref: '#/components/schemas/InspectionLabelEnum'
          readOnly: true
        next_inspection_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        brand:
          type: string
          readOnly: true
        model:
          type: string
          readOnly: true
      required:
      - brand
      - building_id
      - building_name
      - category
      - customer_name
      - has_car_door
      - id
      - inspection_label
      - model
      - name
      - next_inspection_date
      - registration_number
      - status
      - stop_count
    ElevatorStatus:
      enum:
      - active
      - suspended
      - sealed
      - out_of_service
      - uncontracted
      type: string
      description: |-
        * `active` - Active
        * `suspended` - Suspended
        * `sealed` - Sealed
        * `out_of_service` - Out of service
        * `uncontracted` - Uncontracted
    ElevatorWrite:
      type: object
      properties:
        building:
          type: string
          format: uuid
        registration_number:
          type: string
          maxLength: 30
        internal_code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        category:
          oneOf:
          - $ref: '#/components/schemas/ElevatorCategory'
          - $ref: '#/components/schemas/BlankEnum'
        drive_type:
          oneOf:
          - $ref: '#/components/schemas/DriveTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        control_type:
          oneOf:
          - $ref: '#/components/schemas/ControlTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        door_type:
          oneOf:
          - $ref: '#/components/schemas/DoorTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        has_car_door:
          type: boolean
          nullable: true
        machine_room:
          oneOf:
          - $ref: '#/components/schemas/MachineRoomEnum'
          - $ref: '#/components/schemas/BlankEnum'
        capacity_kg:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        capacity_persons:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        stop_count:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        entrance_count:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        speed_mps:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,2})?$
          nullable: true
        pit_depth_mm:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        headroom_mm:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        car_weight_kg:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        brand:
          type: string
          maxLength: 100
        model:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        manufacturer:
          type: string
          maxLength: 150
        installer:
          type: string
          maxLength: 150
        installation_date:
          type: string
          format: date
          nullable: true
        commissioning_date:
          type: string
          format: date
          nullable: true
        ce_certificate_number:
          type: string
          maxLength: 60
        warranty_end_date:
          type: string
          format: date
          nullable: true
        last_inspection_date:
          type: string
          format: date
          nullable: true
        inspection_label:
          allOf:
          - $ref: '#/components/schemas/InspectionLabelEnum'
          default: none
        next_inspection_date:
          type: string
          format: date
          nullable: true
        inspection_body:
          type: string
          maxLength: 150
        inspection_report_number:
          type: string
          maxLength: 60
        status:
          $ref: '#/components/schemas/ElevatorWriteStatusEnum'
        maintenance_interval_days:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
        notes:
          type: string
      required:
      - building
    ElevatorWriteRequest:
      type: object
      properties:
        building:
          type: string
          format: uuid
        registration_number:
          type: string
          maxLength: 30
        internal_code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        category:
          oneOf:
          - $ref: '#/components/schemas/ElevatorCategory'
          - $ref: '#/components/schemas/BlankEnum'
        drive_type:
          oneOf:
          - $ref: '#/components/schemas/DriveTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        control_type:
          oneOf:
          - $ref: '#/components/schemas/ControlTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        door_type:
          oneOf:
          - $ref: '#/components/schemas/DoorTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        has_car_door:
          type: boolean
          nullable: true
        machine_room:
          oneOf:
          - $ref: '#/components/schemas/MachineRoomEnum'
          - $ref: '#/components/schemas/BlankEnum'
        capacity_kg:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        capacity_persons:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        stop_count:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        entrance_count:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        speed_mps:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,2})?$
          nullable: true
        pit_depth_mm:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        headroom_mm:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        car_weight_kg:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        brand:
          type: string
          maxLength: 100
        model:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        manufacturer:
          type: string
          maxLength: 150
        installer:
          type: string
          maxLength: 150
        installation_date:
          type: string
          format: date
          nullable: true
        commissioning_date:
          type: string
          format: date
          nullable: true
        ce_certificate_number:
          type: string
          maxLength: 60
        warranty_end_date:
          type: string
          format: date
          nullable: true
        last_inspection_date:
          type: string
          format: date
          nullable: true
        inspection_label:
          allOf:
          - $ref: '#/components/schemas/InspectionLabelEnum'
          default: none
        next_inspection_date:
          type: string
          format: date
          nullable: true
        inspection_body:
          type: string
          maxLength: 150
        inspection_report_number:
          type: string
          maxLength: 60
        status:
          $ref: '#/components/schemas/ElevatorWriteStatusEnum'
        maintenance_interval_days:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
        notes:
          type: string
      required:
      - building
    ElevatorWriteStatusEnum:
      enum:
      - active
      - suspended
      - sealed
      - out_of_service
      type: string
      description: |-
        * `active` - active
        * `suspended` - suspended
        * `sealed` - sealed
        * `out_of_service` - out_of_service
...
    InspectionLabelEnum:
      enum:
      - green
      - blue
      - yellow
      - red
      - none
      type: string
      description: |-
        * `green` - Green
        * `blue` - Blue
        * `yellow` - Yellow
        * `red` - Red
        * `none` - None
...
    MachineRoomEnum:
      enum:
      - present
      - absent
      - partial
      type: string
      description: |-
        * `present` - Present
        * `absent` - Absent
        * `partial` - Partial
...
    PaginatedElevatorListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ElevatorList'
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT